date is locale not post, also post settings title
